Modify server.c to polling structure
Uses custom PollingData struct to help setup a list of sockets we would
like to connect to including the listener. This will allow for a much
simpler main loop where the `poll` call can be used to determine which
sockets need to process information and act accordingly.

The main processing loop still needs to be written to actually allow for
handling of connections. In the future this can use a dispatch table
to handle different 'endpoints' that a node offers.